# job-api-project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portail Emploi RESTful",
        "description": "API pour la recherche et la gestion d'offres d'emploi, avec gestion des rôles (Admin, Employeur, Utilisateur).",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur API Principal"
        }
    ],
    "paths": {
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Liste tous les utilisateurs (Rôle : admin requis)",
                "operationId": "adminGetUsers",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs."
                    },
                    "403": {
                        "description": "Accès refusé."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{user}": {
            "get": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Affiche les détails d'un utilisateur (Rôle : admin requis)",
                "operationId": "adminShowUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'utilisateur."
                    },
                    "403": {
                        "description": "Accès refusé."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Met à jour les informations d'un utilisateur (Rôle : admin requis)",
                "operationId": "adminUpdateUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour."
                    },
                    "403": {
                        "description": "Accès refusé."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Supprime un utilisateur (Rôle : admin requis)",
                "operationId": "adminDeleteUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Utilisateur supprimé."
                    },
                    "403": {
                        "description": "Accès refusé (impossible de supprimer un admin)."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{user}/roles": {
            "put": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Met à jour les rôles d'un utilisateur (Rôle : admin requis)",
                "operationId": "adminUpdateUserRoles",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roles"
                                ],
                                "properties": {
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "admin",
                                                "employer",
                                                "user"
                                            ]
                                        },
                                        "example": [
                                            "employer",
                                            "user"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rôles mis à jour."
                    },
                    "403": {
                        "description": "Accès refusé."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Personal Access Token",
                "scheme": "bearer"
            }
        }
    },
...
}